requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
groq
python-dotenv
tenacity==8.2.3
